Split up space show into multiple reusable components
