+++ client/src/pages/space-show/space-show.jsx
@@ -1,55 +1,31 @@
 import React, { useEffect } from 'react'
 import PicturesGrid from '../../components/spaces/pictures-grid.jsx'
+import SpaceInfo from "../../components/spaces/space-info"
 import Spinner from '../../components/spinner/spinner'
-import ReviewsContainer from '../../components/reviews/reviews_container'
+import ReserveWidget from '../../components/spaces/reserve-widget'
 
-const SpaceShow = (props) => {
+const SpaceShow = ({space, match, fetchSpace}) => {
     useEffect(() => {
-        const { id: spaceId } = props.match.params
-        props.fetchSpace(spaceId)
+        const { id: spaceId } = match.params
+        fetchSpace(spaceId)
     }, [])
 
-    return props.space ?  (
+    return space ?  (
         <div className="space-show">
-            <PicturesGrid photos={props.space.spacePhotos}/>
+            <PicturesGrid photos={space.spacePhotos}/>
             <div className="space-info-wrapper">
-                <div className="space-show-info">
-                {/* 
-                <div className="space-summary">
-                        <span className="space-show-name">{space.name}</span>
-                        <div className="space-show-city">{space.city}</div>
-                    </div>
-                    <div className="description-wrapper">
-                        <p className="space-show-description">
-                            {space.description}
-                        </p>
-                    </div>
-                    <div className="amenities-wrapper">
-                        <div className="amenities-header">Amenities</div>
-                        <div className="amenities-div">
-                            {space.amenityCategories ? includedAmenities(space.amenityCategories) : `${space.name} has not listed any amenities yet`}
-                        </div>
-                        {space.amenityCategories ? otherAmenities(space.amenityCategories) : null}
-                    </div>
-                */}
-                    <ReviewsContainer />
-                </div>
-                <div className="reserve-widget-wrapper">
-                {/* 
-                    <div className="reserve-widget">
-                        <span className="reserve-cost">${space.cost} </span>
-                        <span className="reserve-per-month">per desk / month</span>
-                    </div>
-                    <div className="contact-info-wrapper">
-                        <div className="contact-item-top">
-                            <a className="space-website-link" href={space.website}>{space.website}</a>
-                        </div>
-                        <div className="contact-item">{space.email}</div>
-                        <div className="contact-item">{space.phone}</div>
-                    </div>
-                */}
- 
-                </div>
+                <SpaceInfo 
+                    name={space.name} 
+                    city={space.city}
+                    description={space.description}
+                    amenityCategories={space.amenityCategories}
+                />
+                <ReserveWidget 
+                    cost={space.cost}
+                    website={space.website}
+                    email={space.email}
+                    phone={space.phone}
+                />
             </div>
         </div>
     ) : (
