+++ client/src/sass/components/_reserve-widget.scss
@@ -1,0 +1,57 @@
+.reserve-widget-wrapper {
+    margin-right: auto;
+    margin-left: 30px;
+    width: 376px;
+    padding-right: 24px;
+}
+
+.reserve-widget {
+    margin-top: 24px;
+    padding-left: 24px;
+    padding-right: 24px;
+    padding-bottom: 12px;
+    padding-top: 12px;
+    border: 1px solid #e6e6e6
+}
+
+.space-website-link {
+    text-decoration: none;
+    color: #4abde9;
+}
+
+.contact-info-wrapper {
+    margin-top: 15px;
+    margin-bottom: 15px;
+    padding-left: 24px;
+    padding-right: 24px;
+
+    padding-top: 12px;
+    border: 1px solid #e6e6e6;
+
+}
+
+.contact-item-top {
+    padding-bottom: 12px;
+}
+.contact-item {
+    border-top: 1px solid #e6e6e6;
+    padding-top: 12px;
+    padding-bottom: 12px;
+}
+
+.reserve-tour-btn {
+    margin-top: 20px;
+    width: 100%;
+    color: white;
+    background-color: #4abde9;
+    font-family: inherit;
+    border-radius: 4px;
+    font-weight: 500;
+    padding: 10px 22px;
+}
+
+.reserve-cost {
+    font-size: 22px;
+    font-weight: 700;
+    max-width: 30%;
+}
