+++ client/src/sass/main.scss
@@ -7,10 +7,10 @@
 @import './components/menu_dropdown';
 @import './components/modal_background';
 @import './components/pictures-grid';
-
+@import './components/reserve-widget';
 @import './components/reviews';
-
 @import './components/session_form';
+@import './components/space-info';
 @import './components/spaces_item';
 @import './components/spaces_wrapper';
 @import './components/_spaces-map';
