+++ client/src/sass/pages/_space-show-page.scss
@@ -11,40 +11,6 @@
     margin-right: auto;
 }
 
-.space-show-info {
-    max-width: 35%;
-    width: 50%;
-    margin-left: auto;
-    margin-right: 30px;
-    width: auto;
-    padding-left: 24px;
-}
-
-
-.space-show-name {  
-    font-size: 32px;
-    font-weight: 600;
-}
-
-.space-show-city {
-    padding-top: 10px;
-    padding-bottom: 10px;
-}
-
-.description-wrapper {
-    border-top: 1px solid #e6e6e6;
-    border-bottom: 1px solid #e6e6e6;
-}
-
-.space-show-description {
-    padding-top: 16px;
-    padding-bottom: 16px;
-}
-
-.space-website-link {
-    text-decoration: none;
-    color: #4abde9;
-}
 
 .amenities-wrapper {
     padding-top: 25px;
@@ -53,13 +19,9 @@
     margin-bottom: 15px;
 }
 
-
-.amenities-header {
-    font-weight: 800
-}
-
-.amenities-div {
-    columns: 2 auto;
+.amenities-modal-wrapper {
+    height: 100vh;
+    width: 100vw;
 }
 
 .amenities-item {
@@ -80,27 +42,6 @@
     color: #4abde9;
 }
 
-
-.space-summary {
-    margin-top: 24px;
-}
-
-.reserve-widget-wrapper {
-    margin-right: auto;
-    margin-left: 30px;
-    width: 376px;
-    padding-right: 24px;
-}
-
-.reserve-widget {
-    margin-top: 24px;
-    padding-left: 24px;
-    padding-right: 24px;
-    padding-bottom: 12px;
-    padding-top: 12px;
-    border: 1px solid #e6e6e6
-}
-
 .carousel-background {
     position: fixed;
     top: 0;
@@ -112,268 +53,15 @@
     z-index: 10;
 }
 
-@media only screen and (max-width: 1050px) {    
-    .main-pic-container {
-        width: 66.66%;
-    }
-    .space-show-space-pics {
-        width: 33.34%;
-    }
-    .space-show-pic-column:first-of-type {
-        display: none;
-    }
-
-    .space-show-pic-column:nth-of-type(2) {
-        width: 100%;
-    }
-}
-
-@media only screen and (max-width: 820px) {
-    .space-info-wrapper {
-        flex-direction: column;
-    }
-
-    .space-show-pic-column {
-        display: none;
-    }
-
-    .space-pics {
-        height: 35vh;
-    }
-
-    .main-pic-container {
-        width: 100%;
-        /* height: 100%; */
-        position: relative;
-    }
-
-    .space-show-main-pic {
-        width: 100%;
-        /* position: relative; */
-        /* height: 100%; */
-    }
-
-    .space-show-space-pics {
-        display: none;
-    }
-
-    .photos-btn {
-      display: none
-    }
-
-    .photos-btn-main-pic {
-        box-shadow: rgba(0, 0, 0, 0.14) 0px 1px 1px 1px !important;
-        background-color: #FFFFFF;
-        padding: 0.5rem 1rem;
-        font-size: 16px;
-        font-family: inherit;
-        font-weight: 500;
-        color: rgb(72, 72, 72);
-        border-radius: 4px;
-        position: absolute !important;
-        left: auto !important;
-        right: 20px !important;
-        bottom: 20px;
-        z-index: 1 !important;
-    }
-
-    .reserve-widget-wrapper {
-        width: 100%;
-        padding-right: 0px;
-        margin-left: 0px;
-        margin-top: 0px;
-        position: sticky;
-        bottom: 0;
-    }
-
-    .reserve-widget {
-        margin-top: 0px;
-        height: auto;
-        z-index: 2;
-        background-color: white;
-        padding-top: 16px !important;
-        padding-bottom: 16px !important;
-        z-index: 4 !important;
-        border-top: 1px solid rgb(235, 235, 235) !important;
-        border-left: 0px;
-        border-right: 0px;
-        border-bottom: 0px;
-    }
-
-    .contact-info-wrapper {
-        display: none;
-    }
-
-    .space-show-info {
-        max-width: 100%;
-    }
-
-    .amenities-wrapper {
-        margin-bottom: 0px;
-    }
-
-    .amenities-modal-wrapper {
-        height: 100vh;
-        width: 100vw;
-    }
-
-    .carousel-background {
-        justify-content: flex-start;
-    }
-}
-
-@media only screen and (max-height: 425px) {
-    .space-info-wrapper {
-        flex-direction: column;
-    }
-
-    .space-show-pic-column {
-        width: 100%;
-    }
-
-    .space-show-pic-column:nth-of-type(2) {
-        display: none;
-    }
-
-    .space-pics {
-        height: 90vh;
-    }
-
-    .main-pic-container {
-        width: 100%;
-        position: relative;
-    }
-
-    .space-show-main-pic {
-        width: 100%;
-        position: relative;
-        height: 100%;
-    }
-
-    .contact-info-wrapper {
-        display: none;
-    }
-
-    .space-show-info {
-        max-width: 100%;
-        margin-left: auto;
-        margin-right: auto;
-    }
-
-    .amenities-wrapper {
-        margin-bottom: 0px;
-    }
-
-    .amenities-modal-wrapper {
-        height: 100vh;
-        width: 100vw;
-    }
-
-    .reserve-widget-wrapper {
-        position: sticky;
-        bottom: 0;
-        background-color: white;
-        width: 100%;
-        padding-right: 0px;
-        margin-left: 0px;
-        margin-top: 0px;
-        z-index: 4;
-        border-top: 1px solid rgb(235, 235, 235) !important;
-
-    }
-
-    .reserve-widget {
-        margin-top: 0px;
-        margin-left: 2rem;
-        margin-right: 2rem; 
-        height: 10vh;
-        /* z-index: 2; */
-        background-color: white;
-        padding-top: 16px !important;
-        padding-bottom: 16px !important;
-        /* z-index: 4 !important; */
-        border-left: 0px;
-        border-right: 0px;
-        border-top: none;
-
-        border-bottom: 0px;
-    }
-
-    .photos-btn {
-        display: none
-    }
-
-    .photos-btn-first-column {
-        display: block;
-        box-shadow: rgba(0, 0, 0, 0.14) 0px 1px 1px 1px !important;
-        background-color: #FFFFFF;
-        padding: 0.5rem 1rem;
-        font-size: 16px;
-        font-family: inherit;
-        font-weight: 500;
-        color: rgb(72, 72, 72);
-        border-radius: 4px;
-        position: absolute !important;
-        left: auto !important;
-        right: 20px !important;
-        bottom: 20px;
-        z-index: 1 !important;
-    }
 
 
-}
-
-.reserve-cost {
-    font-size: 22px;
-    font-weight: 700;
-    max-width: 30%;
-}
 
 
-.reserve-per-month-mobile {
-    font-size: 2em;
-    font-weight: 525;
-    color: rgb(72, 72, 72) !important;
-}
 
-.reserve-tour-btn {
-    margin-top: 20px;
-    width: 100%;
-    color: white;
-    background-color: #4abde9;
-    font-family: inherit;
-    border-radius: 4px;
-    font-weight: 500;
-    padding: 10px 22px;
-}
 
-.contact-info-wrapper {
-    margin-top: 15px;
-    margin-bottom: 15px;
-    padding-left: 24px;
-    padding-right: 24px;
 
-    padding-top: 12px;
-    border: 1px solid #e6e6e6;
 
-}
 
-.contact-info-wrapper-mobile {
-    margin-top: 50px;
-    margin-bottom: 25%;
-    height: 10vh;
-    font-size: 2.75em;
-
-}
-
-.contact-item-top {
-    padding-bottom: 12px;
-}
-.contact-item {
-    border-top: 1px solid #e6e6e6;
-    padding-top: 12px;
-    padding-bottom: 12px;
-}
